split v1 stuff into asm_v1 module
